requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pyyaml
python-dotenv
aiosqlite
httpx
